Refactor register form logic and add step constants
Moved initial register form values, register steps, and DTO transformation logic out of RegisterDTO into dedicated files for better separation of concerns. Updated imports throughout the codebase to use the new locations. Added new dependencies for multistep forms and form handling in package.json.
