+++ Frontend/docs/ACTUAL_STRUCTURE.MD
@@ -1,0 +1,72 @@
+```
+👣📁 src
+    👣📁 assets
+        👣📁 img
+            👣📄 chinacotaxd.jpg
+            👣📄 HarmonyImgNueva.png
+            👣📄 HarmonySocial.png
+            👣📄 imgmusica.jpg
+            👣📄 profile-generig-img-1.png
+            👣📄 profile-generig-img-2.png
+            👣📄 profile-generig-img-3.png
+            👣📄 profile-generig-img-4.png
+            👣📄 profile-generig-img-5.png
+            👣📄 profile-generig-img-6.png
+            👣📄 profile-generig-img-7.png
+            👣📄 profile-generig-img-8.png
+            👣📄 yoxd.jpg
+        👣📁 style
+            👣📄 colors.ts
+        👣📄 defaultAvatars.ts
+    👣📁 components
+        👣📁 auth
+            👣📄 BasicData.tsx
+            👣📄 FavoriteInstrument.tsx
+            👣📄 ProfileImageStep.tsx
+        👣📁 general
+            👣📄 GeneralHeader.tsx
+            👣📄 ProfileImage.tsx
+            👣📄 StepIndicator.tsx
+            👣📄 StepTransition.tsx
+        👣📁 home
+            👣📄 HomeHeaderComponent.tsx
+            👣📄 PostCard.tsx
+    👣📁 config
+        👣📄 AppConfig.ts
+    👣📁 core
+        👣📁 dtos
+            👣📄 RegisterDTO.ts
+            👣📄 RegisterFormData.ts
+        👣📁 hooks
+        👣📁 models
+            👣📄 Playlist.ts
+            👣📄 Post.ts
+            👣📄 User.ts
+        👣📁 services
+            👣📁 general
+                👣📄 ApiService.ts
+            👣📁 user
+                👣📁 auth
+                    👣📄 AuthUserService.ts
+        👣📁 types
+            👣📄 colorsContainer.ts
+            👣📄 RegisterStep.ts
+            👣📄 StepValidation.ts
+    👣📁 navigation
+        👣📄 MainTabs.tsx
+    👣📁 screens
+        👣📁 auth
+            👣📄 LoginScreen.tsx
+            👣📄 ProfileScreen.tsx
+            👣📄 RegisterScreen.tsx
+            👣📄 ResetPasswordScreen.tsx
+        👣📁 home
+            👣📄 HomeScreen.tsx
+    👣📁 utils
+    👣📁 viewmodels
+        👣📁 types
+            👣📄 RegisterViewModelTypes.ts
+            👣📄 stepValidationSchemas.ts
+        👣📄 useRegisterViewModel.ts
+    👣📄 App.tsx
+```
