+++ Frontend/package.json
@@ -10,6 +10,7 @@
     "test": "jest"
   },
   "dependencies": {
+    "@brijen/react-native-multistep": "^1.0.0",
     "@react-native-async-storage/async-storage": "2.2.0",
     "@react-navigation/bottom-tabs": "7.4.6",
     "@react-navigation/native": "7.1.17",
@@ -17,12 +18,14 @@
     "axios": "^1.12.1",
     "formik": "2.4.6",
     "react": "19.1.0",
+    "react-hook-form": "^7.62.0",
     "react-native": "0.80.1",
     "react-native-gesture-handler": "2.28.0",
     "react-native-linear-gradient": "2.8.3",
     "react-native-reanimated": "4.0.2",
     "react-native-safe-area-context": "5.6.0",
     "react-native-screens": "4.14.1",
+    "react-native-svg": "^15.13.0",
     "react-native-vector-icons": "10.3.0",
     "react-native-worklets": "0.4.1",
     "yup": "1.7.0"
