+++ Frontend/src/config/const/InitialRegisterForm.ts
@@ -1,0 +1,15 @@
+import { RegisterFormData } from '../../core/dtos/RegisterFormData';
+
+/**
+ * Valores iniciales para el formulario de registro
+ */
+
+export const INITIAL_REGISTER_FORM: RegisterFormData = {
+  fullName: '',
+  username: '',
+  email: '',
+  password: '',
+  confirmPassword: '',
+  favoriteInstrument: null,
+  profileImage: '',
+};
