+++ Frontend/src/core/utils/transforms/transformToRegisterDTO.ts
@@ -1,0 +1,23 @@
+import { RegisterDTO } from '../../dtos/RegisterDTO';
+import { RegisterFormData } from '../../dtos/RegisterFormData';
+
+/**
+ * Transforma los datos del formulario al DTO esperado por el backend
+ */
+
+export const transformToRegisterDTO = (
+  formData: RegisterFormData,
+): RegisterDTO => {
+  if (!formData.favoriteInstrument) {
+    throw new Error('Favorite instrument is required');
+  }
+
+  return {
+    full_name: formData.fullName.trim(),
+    email: formData.email.trim().toLowerCase(),
+    username: formData.username.trim(),
+    password: formData.password,
+    profile_image: formData.profileImage,
+    favorite_instrument: formData.favoriteInstrument,
+  };
+};
